+++ src/typeguards.ts
@@ -7,7 +7,11 @@ import type { DocumentType, Ref, RefType } from './types';
  * Check if the given document is populated
  * @param doc The Ref with uncertain type
  */
-export function isDocument<T, S extends RefType>(doc: Ref<T, S>): doc is DocumentType<T> {
+export function isDocument<T, S extends RefType>(
+  doc: Ref<T, S>
+  // handle type case of T being "DocumentType" already and being a OR of other types (like for count hooks)
+  // i am not a typescript wizard, so i dont know how to handle this better, this will need to be updated for #730 and #587
+): doc is T extends DocumentType<infer T1, infer T2> ? DocumentType<T1, T2> : T extends object ? DocumentType<T> : never {
   return doc instanceof mongoose.Model;
 }